fix: إصلاح BigInt وإضافة أدوات فحص قاعدة البيانات
- إصلاح خطأ BigInt في /api/test-db
- إضافة RUN_MIGRATIONS_NOW.md دليل تشغيل migrations
- إضافة check-db-status.js لفحص حالة قاعدة البيانات
- المشكلة: الجداول غير موجودة وتحتاج migrations

+++ app/api/test-db/route.ts
@@ -8,7 +8,7 @@ export async function GET() {
     
     // عد الجداول الموجودة
     const tableCount = await prisma.$queryRaw`
-      SELECT COUNT(*) as count 
+      SELECT COUNT(*)::int as count 
       FROM information_schema.tables 
       WHERE table_schema = 'public'
     `;
